actual first draft

+++ layers/com.schoolyear.rstudio/install_scripts/configure_taskbar_layout.ps1
@@ -8,11 +8,58 @@ $scriptLogPrefix = "Taskbar layout"
 $shortcutSourcePath = ".\rstudio_resources\RStudio.lnk"
 $startMenuProgramsPath = Join-Path $env:ProgramData "Microsoft\Windows\Start Menu\Programs"
 $shortcutDestinationPath = Join-Path $startMenuProgramsPath "RStudio.lnk"
+$shortcutLayoutPath = "%ALLUSERSPROFILE%\Microsoft\Windows\Start Menu\Programs\RStudio.lnk"
 $layoutSourcePath = ".\rstudio_resources\taskbar\TaskbarLayoutModification.xml"
 $layoutDirectory = "C:\Windows\OEM"
 $layoutDestinationPath = Join-Path $layoutDirectory "TaskbarLayoutModification.xml"
 $explorerRegistryPath = "HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\Explorer"
 $layoutRegistryName = "LayoutXMLPath"
+
+function Add-RStudioTaskbarPin {
+  param (
+    [Parameter(Mandatory = $true)]
+    [string]$LayoutPath,
+
+    [Parameter(Mandatory = $true)]
+    [string]$DesktopApplicationLinkPath
+  )
+
+  [xml]$layoutXml = Get-Content -LiteralPath $LayoutPath
+
+  $layoutNamespaceUri = "http://schemas.microsoft.com/Start/2014/LayoutModification"
+  $defaultLayoutNamespaceUri = "http://schemas.microsoft.com/Start/2014/FullDefaultLayout"
+  $taskbarNamespaceUri = "http://schemas.microsoft.com/Start/2014/TaskbarLayout"
+
+  $namespaceManager = New-Object System.Xml.XmlNamespaceManager($layoutXml.NameTable)
+  $namespaceManager.AddNamespace("layout", $layoutNamespaceUri)
+  $namespaceManager.AddNamespace("defaultlayout", $defaultLayoutNamespaceUri)
+  $namespaceManager.AddNamespace("taskbar", $taskbarNamespaceUri)
+
+  $pinList = $layoutXml.SelectSingleNode("//taskbar:TaskbarPinList", $namespaceManager)
+
+  if ($null -eq $pinList) {
+    $customTaskbarLayoutCollection = $layoutXml.SelectSingleNode("//layout:CustomTaskbarLayoutCollection", $namespaceManager)
+    if ($null -eq $customTaskbarLayoutCollection) {
+      throw "CustomTaskbarLayoutCollection was not found in $LayoutPath"
+    }
+
+    $taskbarLayout = $layoutXml.SelectSingleNode("//defaultlayout:TaskbarLayout", $namespaceManager)
+    if ($null -eq $taskbarLayout) {
+      $taskbarLayout = $layoutXml.CreateElement("defaultlayout", "TaskbarLayout", $defaultLayoutNamespaceUri)
+      $customTaskbarLayoutCollection.AppendChild($taskbarLayout) | Out-Null
+    }
+
+    $pinList = $layoutXml.CreateElement("taskbar", "TaskbarPinList", $taskbarNamespaceUri)
+    $taskbarLayout.AppendChild($pinList) | Out-Null
+  }
+
+  $desktopApp = $layoutXml.CreateElement("taskbar", "DesktopApp", $taskbarNamespaceUri)
+  $desktopApp.SetAttribute("DesktopApplicationLinkPath", $DesktopApplicationLinkPath)
+  $pinList.AppendChild($desktopApp) | Out-Null
+
+  $layoutXml.Save($LayoutPath)
+  Write-Host "${scriptLogPrefix}: Added RStudio taskbar pin to $LayoutPath"
+}
 
 if (!(Test-Path $shortcutSourcePath)) {
   throw "RStudio shortcut resource was not found at $shortcutSourcePath"
@@ -29,17 +76,22 @@ if (!(Test-Path $shortcutDestinationPath)) {
   throw "RStudio Start Menu shortcut was not found at $shortcutDestinationPath after copying"
 }
 
-if (!(Test-Path $layoutSourcePath)) {
-  throw "Taskbar layout resource was not found at $layoutSourcePath"
-}
-
 if (!(Test-Path $layoutDirectory)) {
   Write-Host "${scriptLogPrefix}: Creating $layoutDirectory"
   New-Item -Path $layoutDirectory -ItemType Directory -Force | Out-Null
 }
 
-Write-Host "${scriptLogPrefix}: Copying taskbar layout to $layoutDestinationPath"
-Copy-Item -Path $layoutSourcePath -Destination $layoutDestinationPath -Force | Out-Null
+if (Test-Path $layoutDestinationPath) {
+  Write-Host "${scriptLogPrefix}: Updating existing taskbar layout at $layoutDestinationPath"
+  Add-RStudioTaskbarPin -LayoutPath $layoutDestinationPath -DesktopApplicationLinkPath $shortcutLayoutPath
+} else {
+  if (!(Test-Path $layoutSourcePath)) {
+    throw "Taskbar layout resource was not found at $layoutSourcePath"
+  }
+
+  Write-Host "${scriptLogPrefix}: Copying taskbar layout to $layoutDestinationPath"
+  Copy-Item -Path $layoutSourcePath -Destination $layoutDestinationPath -Force | Out-Null
+}
 
 if (!(Test-Path $layoutDestinationPath)) {
   throw "Taskbar layout file was not found at $layoutDestinationPath after copying"
